// src/interfaces/user.interface.ts
export interface UserJson {
  id?: string;
  first_name?: string;
  last_name?: string;
  fullname?: string;
  email?: string;
  phone_indicative?: string;
  phone?: string;
  date_joined?: string;
  is_block?: boolean;
  preference_notification?: string;
  last_login?: string;
  user_app_id?: string;
  referral_code?: string;
  referral_link?: string;
  access?: string;
  refresh?: string;
}